Remove methods from enigma class
